Speed up differential equation projection
This change speeds up the projection step in the differential equation
solver by telling the integrator that it does not change the solution.
While this is not true, the change flag is mostly ment for discontinuous
changes in `u`, while the normalisation at each time step should be
continuous within the error tolerances.