Fix errors with newer versions of DifferentialEquations.jl
Newer versions of DifferentialEquations.jl (and its dependency
RecursiveArrayTools.jl) changed the way the solution can be accesed as
an array, causing the example run files to error out. Apparently, the
way the solution was used as an array, although it worked fine before,
it was never officially supported (see:
https://github.com/SciML/DifferentialEquations.jl/issues/1021). This
commit fixed this issue, using the syntax officially supported. This
makes the code work with both the newer versions and the older ones as
well.